backend/requirements.txt
Django==3.1.4

# Python-PostgreSQL Database Adapter
psycopg2-binary==2.8

# Graphql
graphene-django==2.9.1
graphene==2.1.8
django-graphql-jwt==0.3.1

# Custom requirements
django-cors-headers==3.5.0
